fixing some int references and segfaults at runtime

+++ demos/hackrf_tcp_server/include/flex_util.hpp
@@ -5,7 +5,7 @@
 #include <string>
 #include "../../../tinyflex.h"
 
-inline bool encode_flex_message(const std::string& message, int capcode, uint8_t* flex_buffer, size_t flex_buffer_size, size_t& flex_len, int& error) {
+inline bool encode_flex_message(const std::string& message, uint64_t capcode, uint8_t* flex_buffer, size_t flex_buffer_size, size_t& flex_len, int& error) {
     memset(flex_buffer, 0, flex_buffer_size);
     error = 0;
     flex_len = tf_encode_flex_message(message.c_str(), capcode, flex_buffer, flex_buffer_size, &error);
